refactor: extract NotificationGrouper from InboxController
Split the pure grouping/filtering/threading/mention/decryption logic out
of InboxController into a NotificationGrouper. The controller now owns
only state, pagination, sync polling, and mark-as-read, and delegates
grouping (and the isMention / threadRootIdFor / decryptedContentFor
queries) to the grouper.

Behavior-preserving: the grouper takes the active filter as a parameter
and owns the decrypted-content cache (cleared on account switch). Public
API is unchanged; InboxController re-exports InboxFilter and the group
models so existing imports keep working.

Refs #625
